// src/app/_enums/event-type.enum.ts
/**
 * event type
 *
 * names for timeSpan category are assigned via timeSpanCategory option
 *
 */
export enum EventType {
  plan = 0,
  actual = 1,
  timeSpanUndefined = 20,
  timeSpanCategory1 = 21,
  timeSpanCategory2 = 22,
  timeSpanCategory3 = 23,
  timeSpanCategory4 = 24,
  timeSpanCategory5 = 25,
  timeSpanCategory6 = 26,
  timeSpanCategory7 = 27,
  timeSpanCategory8 = 28,
  timeSpanCategory9 = 29
}